refactor(plugins): extract default terminals into side-effect-free terminal.ts
Each default plugin's register.ts ran a module-load registerPlugin() side
effect, so importing a terminal function at a call site (loop.ts,
conversation-agent-loop.ts, etc.) dragged in registration as well. Move each
terminal function (plus its constants and helpers) into a sibling terminal.ts
that has zero side effects, keep registration + the passthrough middleware in
register.ts, and repoint all call sites to import from terminal.ts. This
decouples the terminal computation from plugin registration without changing
any behavior or the middleware chain.
