+++ assistant/src/__tests__/tool-result-truncate-pipeline.test.ts
@@ -24,10 +24,8 @@ import {
   TRUNCATION_SUFFIX,
 } from "../context/tool-result-truncation.js";
 import type { TrustContext } from "../daemon/trust-context.js";
-import {
-  defaultToolResultTruncatePlugin,
-  defaultToolResultTruncateTerminal,
-} from "../plugins/defaults/tool-result-truncate/register.js";
+import { defaultToolResultTruncatePlugin } from "../plugins/defaults/tool-result-truncate/register.js";
+import { defaultToolResultTruncateTerminal } from "../plugins/defaults/tool-result-truncate/terminal.js";
 import { DEFAULT_TIMEOUTS, runPipeline } from "../plugins/pipeline.js";
 import {
   getMiddlewaresFor,
