+++ assistant/src/plugins/defaults/compaction/terminal.ts
@@ -1,0 +1,85 @@
+/**
+ * Terminal handler for the default `compaction` pipeline.
+ *
+ * This module is side-effect free: importing it does not register any plugin.
+ * The terminal is wired in as the pipeline's `terminal` argument by the
+ * `runPipeline` call site in `daemon/conversation-agent-loop.ts`.
+ *
+ * Delegates to the orchestrator's existing
+ * {@link ContextWindowManager} instance. No behavior change relative to the
+ * pre-plugin call site — the plugin only exists so custom plugins registered
+ * in later PRs can observe arguments, short-circuit to a different summary, or
+ * post-process the {@link ContextWindowResult} before the orchestrator
+ * consumes it.
+ *
+ * Lookup: the terminal reads `ctx.contextWindowManager` from the
+ * {@link TurnContext} as a typed optional field. The orchestrator is
+ * responsible for attaching that handle to the per-turn context it hands to
+ * {@link runPipeline}. If the handle is missing, the terminal throws a
+ * {@link PluginExecutionError} so the bug surfaces with clear attribution
+ * instead of a late `undefined.maybeCompact is not a function`.
+ *
+ * Design doc: `.private/plans/agent-plugin-system.md` (PR 25).
+ */
+
+import type {
+  ContextWindowCompactOptions,
+  ContextWindowManager,
+  ContextWindowResult,
+} from "../../../context/window-manager.js";
+import type { Message } from "../../../providers/types.js";
+import {
+  type CompactionArgs,
+  type CompactionResult,
+  PluginExecutionError,
+  type TurnContext,
+} from "../../types.js";
+
+/**
+ * Name under which the default plugin registers. Exposed so tests and later
+ * plugins can assert registration order or override the default via
+ * composition.
+ */
+export const DEFAULT_COMPACTION_PLUGIN_NAME = "default-compaction";
+
+/**
+ * Read `contextWindowManager` off the turn context. Throws
+ * {@link PluginExecutionError} when absent so the failure attributes cleanly
+ * to the default plugin instead of manifesting as a later NPE.
+ */
+function extractManager(ctx: TurnContext): ContextWindowManager {
+  const manager = ctx.contextWindowManager;
+  if (
+    manager == null ||
+    typeof manager !== "object" ||
+    typeof (manager as { maybeCompact?: unknown }).maybeCompact !== "function"
+  ) {
+    throw new PluginExecutionError(
+      "default-compaction: ctx.contextWindowManager is missing — orchestrator must attach it before invoking the compaction pipeline",
+      DEFAULT_COMPACTION_PLUGIN_NAME,
+    );
+  }
+  return manager;
+}
+
+/**
+ * Default terminal behavior. Exposed as a standalone function (rather than
+ * inlined in the plugin object) so the orchestrator can pass it directly to
+ * {@link runPipeline} as the terminal handler. Keeping terminal-vs-middleware
+ * separate avoids a wasted `next → terminal` hop when no custom plugin
+ * observes the slot.
+ */
+export async function defaultCompactionTerminal(
+  args: CompactionArgs,
+  ctx: TurnContext,
+): Promise<CompactionResult> {
+  const manager = extractManager(ctx);
+  const messages = args.messages as Message[];
+  const options = args.options as ContextWindowCompactOptions | undefined;
+  const result: ContextWindowResult = await manager.maybeCompact(
+    messages,
+    args.signal,
+    options,
+  );
+  return result;
+}
