+++ assistant/src/plugins/defaults/persistence/terminal.ts
@@ -1,0 +1,87 @@
+/**
+ * Terminal handler for the default `persistence` pipeline.
+ *
+ * This module is side-effect free: importing it does not register any plugin.
+ * The terminal is wired in as the pipeline's `terminal` argument by the
+ * `runPipeline` call sites in `daemon/conversation-agent-loop.ts` and
+ * `daemon/conversation-agent-loop-handlers.ts`.
+ *
+ * The terminal dispatches on the discriminated {@link PersistArgs.op} field:
+ *
+ * - `add`           → {@link addMessage}, optionally followed by
+ *                     {@link syncMessageToDisk} when `args.syncToDisk` is true.
+ * - `reserve`       → {@link reserveMessage} — pre-allocates an empty row
+ *                     for assistant anchor stamping.
+ * - `updateContent` → {@link updateMessageContent} — overwrites an existing
+ *                     row's content (returns `void`, wrapped as
+ *                     `{ op: "updateContent" }`).
+ * - `update`        → {@link updateMessageMetadata} (returns `void`, wrapped
+ *                     as `{ op: "update" }`).
+ * - `delete`        → {@link deleteMessageById} (returns the segment/summary
+ *                     IDs the caller must clean up out-of-band).
+ */
+
+import {
+  addMessage,
+  deleteMessageById,
+  reserveMessage,
+  updateMessageContent,
+  updateMessageMetadata,
+} from "../../../memory/conversation-crud.js";
+import { syncMessageToDisk } from "../../../memory/conversation-disk-view.js";
+import type { PersistArgs, PersistResult } from "../../types.js";
+
+/**
+ * Terminal handler for the `persistence` pipeline. Exported so tests can
+ * verify default behavior directly without going through `runPipeline`, and
+ * so the `daemon/conversation-agent-loop*.ts` call sites can pass it as the
+ * `terminal` argument to `runPipeline`.
+ */
+export async function defaultPersistenceTerminal(
+  args: PersistArgs,
+): Promise<PersistResult> {
+  switch (args.op) {
+    case "add": {
+      const message = await addMessage(
+        args.conversationId,
+        args.role,
+        args.content,
+        {
+          metadata: args.metadata,
+          ...args.addOptions,
+        },
+      );
+      // Sync the just-persisted row to the JSONL disk view when the caller
+      // opted in. The handler that emits tool-result rows sets
+      // `syncToDisk: true` so the disk view stays in lockstep with the DB.
+      if (args.syncToDisk && args.createdAtMs !== undefined) {
+        syncMessageToDisk(args.conversationId, message.id, args.createdAtMs);
+      }
+      return { op: "add", message };
+    }
+    case "reserve": {
+      const message = await reserveMessage(
+        args.conversationId,
+        args.role,
+        args.metadata,
+      );
+      return { op: "reserve", message };
+    }
+    case "updateContent": {
+      updateMessageContent(args.messageId, args.content);
+      return { op: "updateContent" };
+    }
+    case "update": {
+      updateMessageMetadata(args.messageId, args.updates);
+      return { op: "update" };
+    }
+    case "delete": {
+      const deleted = deleteMessageById(args.messageId);
+      return {
+        op: "delete",
+        segmentIds: deleted.segmentIds,
+        deletedSummaryIds: deleted.deletedSummaryIds,
+      };
+    }
+  }
+}
