+++ assistant/src/plugins/defaults/title-generate/register.ts
@@ -1,63 +1,30 @@
 /**
  * Default `titleGenerate` pipeline plugin.
  *
- * The terminal for the `titleGenerate` pipeline. Delegates to
- * {@link queueGenerateConversationTitle}, which schedules title generation
- * as fire-and-forget background work and falls back to a deterministic
- * placeholder on failure.
- *
- * Custom plugins may install middleware that short-circuits this terminal
- * (e.g. a deterministic generator for tests, or an alternative LLM routing
- * policy). When no middleware is installed the pipeline calls this
- * terminal directly and behavior is bit-identical to the pre-plugin code
- * path.
+ * Declares no middleware — the terminal handler in `./terminal.ts` is wired in
+ * as the pipeline's `terminal` argument by the `runPipeline` call site in
+ * `daemon/conversation-agent-loop.ts`. This plugin exists purely to negotiate
+ * the `titleGenerateApi` capability so bootstrap has a record that the
+ * assistant runtime exposes this pipeline.
  *
  * Registered via a side-effect import from
  * `daemon/external-plugins-bootstrap.ts` so it is present in the registry
  * by the time {@link bootstrapPlugins} runs.
  */
 
-import { queueGenerateConversationTitle } from "../../../memory/conversation-title-service.js";
 import { registerPlugin } from "../../registry.js";
-import {
-  type Plugin,
-  PluginExecutionError,
-  type TitleArgs,
-  type TitleResult,
-} from "../../types.js";
+import { type Plugin, PluginExecutionError } from "../../types.js";
 import pkg from "./package.json" with { type: "json" };
-
-/**
- * Invoke the title-generation service with the provided arguments. Used as
- * the terminal handler for the `titleGenerate` pipeline in
- * `conversation-agent-loop.ts`, and re-exported for tests that want to
- * exercise the default directly.
- *
- * Returns an empty result — the service is fire-and-forget and surfaces its
- * output through `onTitleUpdated`.
- */
-export async function defaultTitleGenerateTerminal(
-  args: TitleArgs,
-): Promise<TitleResult> {
-  queueGenerateConversationTitle({
-    conversationId: args.conversationId,
-    provider: args.provider,
-    userMessage: args.userMessage,
-    onTitleUpdated: args.onTitleUpdated,
-  });
-  return {};
-}
 
 /**
  * Default titleGenerate plugin. Declares no middleware — it exists purely
  * to negotiate the `titleGenerateApi` capability so bootstrap has a record
  * that the assistant runtime exposes this pipeline.
  *
- * The terminal is supplied at the call site in
- * `conversation-agent-loop.ts` (see {@link defaultTitleGenerateTerminal})
- * rather than through `middleware.titleGenerate`, because a default
- * middleware would short-circuit user-registered middleware by always
- * running first in onion order. Keeping the terminal outside the
+ * The terminal handler (`./terminal.ts`) is supplied at the call site in
+ * `conversation-agent-loop.ts` rather than through `middleware.titleGenerate`,
+ * because a default middleware would short-circuit user-registered middleware
+ * by always running first in onion order. Keeping the terminal outside the
  * middleware chain lets user plugins observe/transform/short-circuit the
  * call without competing with an assistant-owned default middleware.
  */
