+++ assistant/src/plugins/defaults/tool-error/terminal.ts
@@ -1,0 +1,54 @@
+/**
+ * Terminal handler for the default `toolError` pipeline.
+ *
+ * This module is side-effect free: importing it does not register any plugin.
+ * The terminal is wired in as the pipeline's `terminal` argument by the
+ * `runPipeline` call site in `agent/loop.ts`.
+ *
+ * The canonical nudge decision: when the current turn produced at least one
+ * failed tool result, append a system-notice block to the tool results that
+ * coaches the LLM to either retry with corrected parameters (for recoverable
+ * errors) or report the failure to the user (for unrecoverable ones). Once
+ * the consecutive-error-turn counter exceeds the caller-supplied cap, the
+ * nudge is skipped — the error is likely not something the LLM can fix on
+ * its own and continuing to nudge only burns tokens.
+ *
+ * Design doc: `.private/plans/agent-plugin-system.md` (PR 19).
+ */
+
+import type { ToolErrorArgs, ToolErrorDecision } from "../../types.js";
+
+/**
+ * Canonical nudge text. Kept as a module-level constant so tests and future
+ * plugins can match it without duplicating the string.
+ */
+export const DEFAULT_TOOL_ERROR_NUDGE_TEXT =
+  "<system_notice>One or more tool calls returned an error. If the error looks recoverable (e.g. missing or invalid parameters), fix the parameters and retry. If the error is clearly unrecoverable (e.g. a service is down, a resource does not exist, or a permission is permanently denied), report it to the user.</system_notice>";
+
+/**
+ * Terminal handler for the `toolError` pipeline. Nudge iff the current turn
+ * had an error AND the consecutive-error counter is within the cap. Once the
+ * cap is breached the caller should stop appending the nudge (the error is
+ * likely unrecoverable and the LLM already had multiple attempts to correct
+ * it).
+ *
+ * Exported so `agent/loop.ts` can pass it as the `terminal` argument to
+ * `runPipeline` (ensuring the nudge decision fires even when no plugin is
+ * registered — e.g. direct AgentLoop callers that skip `bootstrapPlugins()`)
+ * and so tests can verify the decision logic directly without going through
+ * the pipeline runner.
+ */
+export const defaultToolErrorTerminal = async (
+  args: ToolErrorArgs,
+): Promise<ToolErrorDecision> => {
+  if (
+    args.hasToolError &&
+    args.consecutiveErrorTurns <= args.maxConsecutiveErrorNudges
+  ) {
+    return {
+      action: "nudge",
+      nudgeText: DEFAULT_TOOL_ERROR_NUDGE_TEXT,
+    };
+  }
+  return { action: "skip" };
+};
